reconstruct native contract
