+++ smartcontract/service/native/governance/states.go
@@ -16,7 +16,7 @@
  * along with The ontology.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package states
+package governance
 
 import (
 	"math/big"
@@ -98,7 +98,7 @@ type Configuration struct {
 	BlockMsgDelay        uint32 `json:"block_msg_delay"`
 	HashMsgDelay         uint32 `json:"hash_msg_delay"`
 	PeerHandshakeTimeout uint32 `json:"peer_handshake_timeout"`
-	MaxBlockChangeView   uint64 `json:"MaxBlockChangeView"`
+	MaxBlockChangeView   uint32 `json:"MaxBlockChangeView"`
 }
 
 type VoteCommitDposParam struct {
